+++ CHANGELOG.md
@@ -15,7 +15,7 @@ Recommendation: for ease of reading, use the following order:
 ### Added
 - GQL: `dataset_metadata`: Added new method `MetadataProjection` to get `[MetadataExtendedBlock]` filtered by types and range
 ### Fixed
-- Correctly handle empty authentication token error
+- Return bad request response for requests with empty bearer authentication token
 
 ## [0.245.3] - 2025-07-29
 ### Fixed